Remove warnings

+++ src/sm4_ccm.c
@@ -159,7 +159,7 @@ int sm4_ccm_decrypt(const SM4_KEY *sm4_key, const uint8_t *iv, size_t ivlen,
 	}
 
 	inlen_size = 15 - ivlen;
-	if (inlen_size < 8 && inlen >= (1 << (inlen_size * 8))) {
+	if (inlen_size < 8 && inlen >= (size_t)(1 << (inlen_size * 8))) {
 		error_print();
 		return -1;
 	}
@@ -181,7 +181,7 @@ int sm4_ccm_decrypt(const SM4_KEY *sm4_key, const uint8_t *iv, size_t ivlen,
 		if (aadlen < ((1<<16) - (1<<8))) {
 			length_to_bytes(aadlen, 2, block);
 			alen = 2;
-		} else if (aadlen < ((size_t)1<<32)) {
+		} else if ((uint64_t)aadlen < ((uint64_t)1<<32)) {
 			block[0] = 0xff;
 			block[1] = 0xfe;
 			length_to_bytes(aadlen, 4, block + 2);
